cost-optimization/gke-vpa-recommendations/gke_metrics_pipeline/1-gke-namespace-metric-builder/requirements.txt
google-cloud-pubsub
httpx
pandas